+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -14,6 +14,8 @@
 - Number of selectable chart elements is now limited to 5.
 - Color scale logic for charts with lots of items has been fixed.
 - Update `@woocommerce/navigation` to v2.0.0
+- Bug fix for `<StockReportTable />` returning N/A instead of zero.
+- In `<Search>` use backspace key to remove tags from the search box.
 
 # 1.4.2
 - Add emoji-flags dependency